move segment constants to separate file
