+++ packages/next/src/server/app-render/create-flight-router-state-from-loader-tree.ts
@@ -1,7 +1,7 @@
 import type { LoaderTree } from '../lib/app-dir-module'
 import type { FlightRouterState, Segment } from './types'
 import type { GetDynamicParamFromSegment } from './app-render'
-import { PAGE_SEGMENT_KEY } from '../../shared/lib/constants'
+import { PAGE_SEGMENT_KEY } from '../../shared/lib/segment'
 
 export function addSearchParamsIfPageSegment(
   segment: Segment,
