+++ packages/next/src/shared/lib/segment.ts
@@ -2,3 +2,6 @@ export function isGroupSegment(segment: string) {
   // Use array[0] for performant purpose
   return segment[0] === '(' && segment.endsWith(')')
 }
+
+export const PAGE_SEGMENT_KEY = '__PAGE__'
+export const DEFAULT_SEGMENT_KEY = '__DEFAULT__'
